Create classes for pre- and post-processing
